优化

+++ springboot/src/main/java/com/sqn/library/controller/ActivityController.java
@@ -40,7 +40,7 @@ public class ActivityController {
     }
 
     @DeleteMapping("/{id}")
-    public Result<?> delete(@PathVariable Integer id) {
+    public Result<?> delete(@PathVariable Long id) {
         activityService.removeById(id);
         return Result.success();
     }
@@ -58,7 +58,7 @@ public class ActivityController {
     }
 
     @GetMapping("/{id}")
-    public Result<?> findOne(@PathVariable Integer id) {
+    public Result<?> findOne(@PathVariable Long id) {
         return Result.success(activityService.getById(id));
     }
